backend/PhotoChallenge/requirements.txt
djangorestframework==3.3.3
